requirements.txt
torch==1.5.1+cpu
torchvision==0.6.1+cpu
numpy==1.21.0
Pillow==8.3.1
-f https://download.pytorch.org/whl/torch_stable.html